wip: rename action sendContactMeGetFormDataAction
